Rename `AllTasks` to `TaskList`
