+++ frontend/src/main.js
@@ -150,7 +150,7 @@ window.addEventListener('DOMContentLoaded', focusTab);
 
 import { defineCustomElement, h } from 'vue';
 import SuspensionWrapper from './components/SuspensionWrapper.vue';
-import AllTasks from './Teacher/AllTasks.vue';
+import TaskList from './Teacher/TaskList.vue';
 import InbusImport from './Teacher/InbusImport.vue';
 import NotificationsNew from './components/Notifications.vue';
 import ColorThemeNew from './components/ColorTheme.vue';
@@ -189,7 +189,7 @@ const registerSuspendedVueComponent = (name, component, configureApp = undefined
     );
 };
 
-registerSuspendedVueComponent('tasks-all', AllTasks);
+registerSuspendedVueComponent('task-list', TaskList);
 registerSuspendedVueComponent('inbus-import', InbusImport);
 registerVueComponent('notifications', NotificationsNew);
 if (enableNewUI) registerVueComponent('color-theme', ColorThemeNew);
