+++ templates/web/task_list.html
@@ -1,3 +1,3 @@
 {% extends 'web/layout.html' %} {% block content %}
-    <kelvin-tasks-all></kelvin-tasks-all>
+    <kelvin-task-list></kelvin-task-list>
 {% endblock %}
