+++ web/urls.py
@@ -133,7 +133,7 @@ urlpatterns = [
     path(
         "task/<int:task_id>/csv", teacher_view.download_csv_per_task, name="download_csv_per_task"
     ),
-    path("tasks", teacher_view.all_tasks, name="tasks"),
+    path("tasks", teacher_view.task_list, name="tasks"),
     path("reevaluate/<int:submit_id>", teacher_view.reevaluate, name="reevaluate"),
     path("api_token", common_view.api_token, name="api_token"),
 ]
